+++ src/basilisp/lang/compiler/generator.py
@@ -1507,7 +1507,7 @@ def __fn_args_to_py_ast(
     fn_args, varg = [], None
     fn_body_ast: List[ast.AST] = []
     for binding in params:
-        assert binding.init is None, ":fn nodes cannot have bindint :inits"
+        assert binding.init is None, ":fn nodes cannot have binding :inits"
         assert varg is None, "Must have at most one variadic arg"
         arg_name = genname(munge(binding.name))
 
@@ -3525,7 +3525,7 @@ def _const_type_to_py_ast(form: IType, ctx: GeneratorContext) -> GeneratedPyAST:
 
     ctor_args = []
     ctor_arg_deps: List[ast.AST] = []
-    for field in attr.fields(tp):
+    for field in attr.fields(tp):  # type: ignore[arg-type]
         field_nodes = _const_val_to_py_ast(getattr(form, field.name, None), ctx)
         ctor_args.append(field_nodes.node)
         ctor_args.extend(field_nodes.dependencies)