+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Added a Sphinx autodoc plugin for generating API documentation for Basilisp namespaces (#658)
  * Added support for rewriting required namespaces starting with `clojure.` as `basilisp.` if the original name isn't found on the import path (#670, #676)
  * Added support for inlining simple functions (#673)
+ * Added the `clojure.core` functions from v1.11 (#672)
 
 ### Changed
  * Set tighter bounds on dependency version ranges (#657)